ArunEstatesTestOlivia/Services/ProductService.cs
using System.Collections.Generic;
using System.Threading.Tasks;
using Microsoft.EntityFrameworkCore;
using ArunEstatesTestOlivia.Data;
using ArunEstatesTestOlivia.Models;

namespace ArunEstatesTestOlivia.Services
{
    public class ProductService : IProductService
    {
        private readonly AppDbContext _context;

        public ProductService(AppDbContext context)
        {
            _context = context;
        }

        public async Task<List<Product>> GetAllProductsAsync()
        {
            return await _context.Products.ToListAsync();
        }

        public async Task<Product?> GetProductByIdAsync(int id)
        {
            return await _context.Products.FindAsync(id);
        }

        public async Task<List<Product>> SearchProductsAsync(string searchTerm)
        {
            return await _context.Products
                .Where(p => p.Name!.Contains(searchTerm) || p.Description!.Contains(searchTerm))
                .ToListAsync();
        }
    }
}
